split out test-facet-retention.js
